Move encoded_permissions tests to separate file.
